refactor(server): peel chat routes into routers/chat.py (router-split, round-5 #51 / R5-25 PR15)
Fourth router peeled (merge-as-green off main after projects/PR14). The /api/chat
group — dispatch a prompt, read a conversation's history, list conversations —
moves into routers/chat.py with the ChatRequest/ChatResponse models (incl. the
empty-prompt validator + project_scope cap) and the conversation-ownership SQL
filter (_chat_owner_filter: remote tokens see only their own conversations,
loopback/admin see all). Imports auth + chat + db — no server import.

_split_csv (search helper) and _bootstrap_admin_token, interleaved in the same
region, deliberately stay in server.py.

test_server.test_chat_persists_capped_prompt referenced server.ChatRequest
directly; repointed to routers.chat.ChatRequest (the model's new home) rather than
re-export models back into server. New boundary test pins leaf/ownership/mount +
that the prompt validator and owner filter travel with the router. 847 passed.
